build: run long-running tests also for drafts
If these tests don't run, codecov/project fails because some code
paths are not covered. When the PR is marked as ready for review,
the tests and codecov are not updated. It's simpler to run these
tests for draft PRs than it would be to re-run the entire test
suite when a PR goes from draft to ready for review.